+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LazyList` now raises `TypeError` if `__len__` is called before the underlying iterable is depleted. This is to
   behave like an iterator to preserve laziness, as certain libraries (e.g., `tqdm`) checks whether an iterable is a list
   and handles things differently.
+- `run_command` and `safe_pool` now uses `log` instead of built-in `print` function.
 
 ## [0.2.0] - 2020-04-20
 ### Added